appBackend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
filelock==3.12.4
Flask==3.0.3
Flask-Cors==4.0.1
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
platformdirs==3.11.0
psycopg2==2.9.9
python-dotenv==1.0.1
virtualenv==20.24.5
Werkzeug==3.0.3
